remove old module
